LPS-52265 Rename
